PGPE: WOV Updates for Actuate Step
CQ: SW529197
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120293
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ import/chips/p10/procedures/ppe/pgpe/pgpe_gppb.c
@@ -880,7 +880,7 @@ uint8_t pgpe_gppb_get_wov_underv_max_pct()
         return G_gppb_wov->wov_underv_max_pct;
     }
 }
-uint8_t pgpe_gppb_get_wov_underv_vmin_mv()
+uint16_t pgpe_gppb_get_wov_underv_vmin_mv()
 {
     if(G_gppb->magic.value == PSTATE_PARMSBLOCK_MAGIC)
     {
@@ -891,7 +891,7 @@ uint8_t pgpe_gppb_get_wov_underv_vmin_mv()
         return G_gppb_wov->wov_underv_vmin_mv;
     }
 }
-uint8_t pgpe_gppb_get_wov_overv_vmax_mv()
+uint16_t pgpe_gppb_get_wov_overv_vmax_mv()
 {
     if(G_gppb->magic.value == PSTATE_PARMSBLOCK_MAGIC)
     {
